Animate elevation when card is picked up.
Had to adjust a tonne of the styling to make this work and look good.

Essentially, on iOS the shadows are just cast across any boundaries of the element, but with Android, you need to provide padding (or margin) for those shadows and scaling to come through.

So, instead of having a horizontal margin on the entire SortableList, I set the horizontal margin on the rows themselves. This gives plenty of room to conduct those animations and shadows.

I also added a similar scaling transform for Android, except a little smaller so we can get more of the shadow from elevation in.

I tested out different elevation ranges and this looks to be about right.